fix tests (again
