+++ api-admin/src/main/java/bio/terra/pearl/api/admin/service/scheduled/ScheduledAssignmentService.java
@@ -1,16 +1,19 @@
 package bio.terra.pearl.api.admin.service.scheduled;
 
-import bio.terra.pearl.core.service.survey.SurveyTaskDispatcher;
+import bio.terra.pearl.core.service.workflow.TaskDispatcher;
+import java.util.List;
 import lombok.extern.slf4j.Slf4j;
 import net.javacrumbs.shedlock.spring.annotation.SchedulerLock;
 import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Component;
 
 @Slf4j
-public class ScheduledSurveyAssignmentService {
-  private final SurveyTaskDispatcher surveyTaskDispatcher;
+@Component
+public class ScheduledAssignmentService {
+  private final List<TaskDispatcher> taskDispatchers;
 
-  public ScheduledSurveyAssignmentService(SurveyTaskDispatcher surveyTaskDispatcher) {
-    this.surveyTaskDispatcher = surveyTaskDispatcher;
+  public ScheduledAssignmentService(List<TaskDispatcher> taskDispatchers) {
+    this.taskDispatchers = taskDispatchers;
   }
 
   @Scheduled(
@@ -21,8 +24,8 @@ public class ScheduledSurveyAssignmentService {
       lockAtMostFor = "500s",
       lockAtLeastFor = "10s")
   public void assignScheduledSurveys() {
-    log.info("Scheduled survey processing beginning");
-    surveyTaskDispatcher.assignScheduledSurveys();
-    log.info("Scheduled survey processing complete");
+    log.info("Scheduled task processing beginning");
+    taskDispatchers.forEach(TaskDispatcher::assignScheduledTasks);
+    log.info("Scheduled task processing complete");
   }
 }
